api/item/delete/route.tsxの修正。引数の型エラーの修正。

+++ app/api/item/delete/[id]/route.tsx
@@ -4,20 +4,27 @@ import { NextRequest, NextResponse } from "next/server";
 
 export async function DELETE(
   request: NextRequest,
-  context: { params: { id: string } }
+  { params }: { params: { id: string } }
 ) {
-  const reqBody = await request.json();
   try {
     await connectDB();
-    const { id } = await context.params;
+    const { id } = await params;
+
     const singleItem = await ItemModel.findById(id);
+
+    if (!singleItem) {
+      return NextResponse.json({ message: "アイテムが見つかりません" });
+    }
+
+    const reqBody = await request.json();
     if (singleItem.email === reqBody.email) {
       await ItemModel.deleteOne({ _id: id });
       return NextResponse.json({ message: "アイテム削除成功" });
     } else {
       return NextResponse.json({ message: "他の人が作成したアイテムです" });
     }
-  } catch {
+  } catch (error) {
+    console.error(error);
     return NextResponse.json({ message: "アイテム削除失敗" });
   }
 }
